BioCrypt-V0.1/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cryptography==44.0.1
customtkinter==5.2.2
cytoolz==1.0.1
darkdetect==0.8.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
mpmath==1.3.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyserial==3.5
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rsa==4.9
sniffio==1.3.1
sympy==1.13.3
tkinterdnd2==0.4.3
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
vaderSentiment==3.3.2
web3==7.8.0
websockets==13.1
yarl==1.18.3
